// src/person.person/dto/create-person.person.dto.ts
export class CreatePersonPersonDto {

    
    PersonType?: String;
    NameStyle?: boolean;
    Title?: String;
    FirstName?: String;
    MiddleName?: String;
    LastName?: String;
    Suffix?: String;
    EmailPromotion?: number;
    AdditionalContactInfo?: String;
    Demographics?: String;
    rowguid?: String;
    ModifiedDate?: Date;
}